Added firebase authentication and login to app
Added firebase authentication to the app.
Added email and password text fields to MainActivity
Added log in button to MainActivity
Added onClick functionality to button
The button's onClick() method authenticates with firebase and logs the
user in
Added ProfileActivity to the project
If login is successful or user is already logged in when app starts, the
app goes to Profile activity where it displays their email address and a
logout button
When the logout button is clicked it logs the user out and returns to
MainActivity which is the login screen.
Code is commented with what all added code does.

I tested this app on my phone and it works.