Added PhotoUploadActvity
This activity will be used to upload the service hour form.
Right now the galery and camera buttons work to select in image from the
galery or take a picture.
Getting the picture to show up once clicked isnt working yet.
Uploading the picture to the firebase server isn't working yet.