Updated PhotoUploadActivity
The camerabutton opens the camera, takes a picture, and returns a
thumbnail above the buttons.
The gallerybutton opens the gallery, selects a picture BUT DOESNT RETURN
THE PICTURE TO THE IMAGE VIEW!!!!!!!!!!
I keep getting a "requires android.permission.READ_EXTERNAL_STORAGE"
error even though that permission is added in the manifest.