Got contact page working, fixed database structure, added app used to fix database
I got the contact page working with a listview. I created an app that
stepped through the database, copied all the users in the old tree, made
user accounts (authentication) for all of them, and created a new tree
in the database for them based on UID. This is also the first commit to
add firebase database connection to the app.